refactor: split storage_schema into separate files
Working towards making this code easier to read/understand.
Small, focused files are easier to maintain.
