+++ twenty-first/src/storage/storage_schema/enums.rs
@@ -1,0 +1,10 @@
+pub enum WriteOperation<K, V> {
+    Write(K, V),
+    Delete(K),
+}
+
+pub enum VecWriteOperation<Index, T> {
+    OverWrite((Index, T)),
+    Push(T),
+    Pop,
+}
