+++ twenty-first/src/storage/storage_schema/mod.rs
@@ -1,968 +1,48 @@
-use super::level_db::DB;
-use super::storage_vec::{Index, StorageVec};
-use itertools::Itertools;
-use leveldb::{
-    batch::{Batch, WriteBatch},
-    options::{ReadOptions, WriteOptions},
-};
-use std::{
-    cell::RefCell,
-    collections::{HashMap, VecDeque},
-    fmt::Debug,
-    sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard},
-};
-
-use crate::shared_math::b_field_element::BFieldElement;
-
-pub type DbtSingletonReference<PK, PV, T> = Arc<RefCell<DbtSingleton<PK, PV, T>>>;
-pub type DbtVecReference<PK, PV, T> = Arc<RefCell<DbtVec<PK, PV, Index, T>>>;
-
-pub enum WriteOperation<ParentKey, ParentValue> {
-    Write(ParentKey, ParentValue),
-    Delete(ParentKey),
-}
-
-pub trait DbTable<ParentKey, ParentValue> {
-    fn pull_queue(&mut self) -> Vec<WriteOperation<ParentKey, ParentValue>>;
-    fn restore_or_new(&mut self);
-}
-
-pub trait StorageReader<ParentKey, ParentValue> {
-    /// Return multiple values from storage, in the same order as the input keys
-    fn get_many(&self, keys: &[ParentKey]) -> Vec<Option<ParentValue>>;
-
-    /// Return a single value from storage
-    fn get(&self, key: ParentKey) -> Option<ParentValue>;
-}
-
-pub enum VecWriteOperation<Index, T> {
-    OverWrite((Index, T)),
-    Push(T),
-    Pop,
-}
-
-pub struct DbtVecPrivate<ParentKey, ParentValue, Index, T> {
-    reader: Arc<dyn StorageReader<ParentKey, ParentValue> + Send + Sync>,
-    current_length: Option<Index>,
-    key_prefix: u8,
-    write_queue: VecDeque<VecWriteOperation<Index, T>>,
-    cache: HashMap<Index, T>,
-    name: String,
-}
-
-pub struct DbtVec<ParentKey, ParentValue, Index, T> {
-    // note: Arc is not needed, because we never hand out inner to anyone.
-    inner: RwLock<DbtVecPrivate<ParentKey, ParentValue, Index, T>>,
-}
-
-impl<ParentKey, ParentValue, T> DbtVec<ParentKey, ParentValue, Index, T>
-where
-    ParentKey: From<(ParentKey, ParentKey)>,
-    ParentKey: From<u8>,
-    ParentKey: From<Index>,
-    Index: From<ParentValue> + From<u64> + Clone,
-    T: Clone,
-{
-    // DbtVec cannot be instantiated directly outside of this crate.
-    pub(crate) fn new(
-        reader: Arc<dyn StorageReader<ParentKey, ParentValue> + Send + Sync>,
-        key_prefix: u8,
-        name: &str,
-    ) -> Self {
-        let vec = DbtVecPrivate::<ParentKey, ParentValue, Index, T>::new(reader, key_prefix, name);
-
-        Self {
-            inner: RwLock::new(vec),
-        }
-    }
-
-    pub(crate) fn read_lock(
-        &self,
-    ) -> RwLockReadGuard<'_, DbtVecPrivate<ParentKey, ParentValue, Index, T>> {
-        self.inner.read().unwrap()
-    }
-
-    pub(crate) fn write_lock(
-        &mut self,
-    ) -> RwLockWriteGuard<'_, DbtVecPrivate<ParentKey, ParentValue, Index, T>> {
-        self.inner.write().unwrap()
-    }
-}
-
-impl<ParentKey, ParentValue, T> StorageVec<T> for DbtVec<ParentKey, ParentValue, Index, T>
-where
-    ParentKey: From<Index>,
-    ParentValue: From<T>,
-    T: Clone + From<ParentValue> + Debug,
-    ParentKey: From<(ParentKey, ParentKey)>,
-    ParentKey: From<u8>,
-    Index: From<ParentValue> + From<u64>,
-{
-    fn is_empty(&self) -> bool {
-        self.read_lock().is_empty()
-    }
-
-    fn len(&self) -> Index {
-        self.read_lock().len()
-    }
-
-    fn get(&self, index: Index) -> T {
-        self.read_lock().get(index)
-    }
-
-    fn get_many(&self, indices: &[Index]) -> Vec<T> {
-        self.read_lock().get_many(indices)
-    }
-
-    fn get_all(&self) -> Vec<T> {
-        self.read_lock().get_all()
-    }
-
-    fn set(&mut self, index: Index, value: T) {
-        self.write_lock().set(index, value)
-    }
-
-    fn set_many(&mut self, key_vals: impl IntoIterator<Item = (Index, T)>) {
-        self.write_lock().set_many(key_vals)
-    }
-
-    fn pop(&mut self) -> Option<T> {
-        self.write_lock().pop()
-    }
-
-    fn push(&mut self, value: T) {
-        self.write_lock().push(value)
-    }
-}
-
-impl<ParentKey, ParentValue, T> DbTable<ParentKey, ParentValue>
-    for DbtVec<ParentKey, ParentValue, Index, T>
-where
-    ParentKey: From<Index>,
-    ParentValue: From<T>,
-    T: Clone,
-    T: From<ParentValue>,
-    ParentKey: From<(ParentKey, ParentKey)>,
-    ParentKey: From<u8>,
-    Index: From<ParentValue>,
-    ParentValue: From<Index>,
-{
-    /// Collect all added elements that have not yet been persisted
-    fn pull_queue(&mut self) -> Vec<WriteOperation<ParentKey, ParentValue>> {
-        self.write_lock().pull_queue()
-    }
-
-    fn restore_or_new(&mut self) {
-        self.write_lock().restore_or_new()
-    }
-}
-
-impl<ParentKey, ParentValue, T> StorageVec<T> for DbtVecReference<ParentKey, ParentValue, T>
-where
-    ParentKey: From<Index>,
-    ParentValue: From<T>,
-    T: Clone + From<ParentValue> + Debug,
-    ParentKey: From<(ParentKey, ParentKey)>,
-    ParentKey: From<u8>,
-    Index: From<ParentValue> + From<u64>,
-{
-    fn is_empty(&self) -> bool {
-        self.borrow().is_empty()
-    }
-
-    fn len(&self) -> Index {
-        self.borrow().len()
-    }
-
-    fn get(&self, index: Index) -> T {
-        self.borrow().get(index)
-    }
-
-    fn get_many(&self, indices: &[Index]) -> Vec<T> {
-        self.borrow().get_many(indices)
-    }
-
-    fn get_all(&self) -> Vec<T> {
-        self.borrow().get_all()
-    }
-
-    fn set(&mut self, index: Index, value: T) {
-        self.borrow_mut().set(index, value)
-    }
-
-    fn set_many(&mut self, key_vals: impl IntoIterator<Item = (Index, T)>) {
-        self.borrow_mut().set_many(key_vals)
-    }
-
-    fn pop(&mut self) -> Option<T> {
-        self.borrow_mut().pop()
-    }
-
-    fn push(&mut self, value: T) {
-        self.borrow_mut().push(value)
-    }
-}
-
-impl<ParentKey, ParentValue, T> DbTable<ParentKey, ParentValue>
-    for DbtVecReference<ParentKey, ParentValue, T>
-where
-    ParentKey: From<Index>,
-    ParentValue: From<T>,
-    T: Clone,
-    T: From<ParentValue>,
-    ParentKey: From<(ParentKey, ParentKey)>,
-    ParentKey: From<u8>,
-    Index: From<ParentValue>,
-    ParentValue: From<Index>,
-{
-    /// Collect all added elements that have not yet been persisted
-    fn pull_queue(&mut self) -> Vec<WriteOperation<ParentKey, ParentValue>> {
-        self.borrow_mut().pull_queue()
-    }
-
-    fn restore_or_new(&mut self) {
-        self.borrow_mut().restore_or_new()
-    }
-}
-
-impl<ParentKey, ParentValue, T> DbtVecPrivate<ParentKey, ParentValue, Index, T>
-where
-    ParentKey: From<(ParentKey, ParentKey)>,
-    ParentKey: From<u8>,
-    ParentKey: From<Index>,
-    Index: From<ParentValue> + From<u64> + Clone,
-    T: Clone,
-{
-    // Return the key of ParentKey type used to store the length of the vector
-    fn get_length_key(key_prefix: u8) -> ParentKey {
-        let const_length_key: ParentKey = 0u8.into();
-        let key_prefix_key: ParentKey = key_prefix.into();
-        (key_prefix_key, const_length_key).into()
-    }
-
-    /// Return the length at the last write to disk
-    fn persisted_length(&self) -> Option<Index> {
-        self.reader
-            .get(Self::get_length_key(self.key_prefix))
-            .map(|v| v.into())
-    }
-
-    /// Return the key of ParentKey type used to store the element at a given index of Index type
-    fn get_index_key(&self, index: Index) -> ParentKey {
-        let key_prefix_key: ParentKey = self.key_prefix.into();
-        let index_key: ParentKey = index.into();
-        (key_prefix_key, index_key).into()
-    }
-
-    pub(crate) fn new(
-        reader: Arc<dyn StorageReader<ParentKey, ParentValue> + Send + Sync>,
-        key_prefix: u8,
-        name: &str,
-    ) -> Self {
-        let length = None;
-        let cache = HashMap::new();
-        Self {
-            key_prefix,
-            reader,
-            write_queue: VecDeque::default(),
-            current_length: length,
-            cache,
-            name: name.to_string(),
-        }
-    }
-
-    fn write_op_overwrite(&mut self, index: Index, value: T) {
-        if let Some(_old_val) = self.cache.insert(index, value.clone()) {
-            // If cache entry exists, we remove any corresponding
-            // OverWrite ops in the `write_queue` to reduce disk IO.
-
-            // logic: retain all ops that are not overwrite, and
-            // overwrite ops that do not have an index matching cache_index.
-            self.write_queue.retain(|op| match op {
-                VecWriteOperation::OverWrite((i, _)) => *i != index,
-                _ => true,
-            })
-        }
-
-        self.write_queue
-            .push_back(VecWriteOperation::OverWrite((index, value)));
-    }
-}
-
-impl<ParentKey, ParentValue, T> StorageVec<T> for DbtVecPrivate<ParentKey, ParentValue, Index, T>
-where
-    ParentKey: From<Index>,
-    ParentValue: From<T>,
-    T: Clone + From<ParentValue> + Debug,
-    ParentKey: From<(ParentKey, ParentKey)>,
-    ParentKey: From<u8>,
-    Index: From<ParentValue> + From<u64>,
-{
-    /// note: acquires lock
-    fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// note: acquires lock
-    fn len(&self) -> Index {
-        self.current_length
-            .unwrap_or_else(|| self.persisted_length().unwrap_or(0))
-    }
-
-    fn get(&self, index: Index) -> T {
-        // Disallow getting values out-of-bounds
-
-        assert!(
-            index < self.len(),
-            "Out-of-bounds. Got {index} but length was {}. persisted vector name: {}",
-            self.len(),
-            self.name
-        );
-
-        // try cache first
-        if self.cache.contains_key(&index) {
-            return self.cache.get(&index).unwrap().clone();
-        }
-
-        // then try persistent storage
-        let key: ParentKey = self.get_index_key(index);
-        let val = self.reader.get(key).unwrap_or_else(|| {
-            panic!(
-                "Element with index {index} does not exist in {}. This should not happen",
-                self.name
-            )
-        });
-        val.into()
-    }
-
-    /// Fetch multiple elements from a `DbtVec` and return the elements matching the order
-    /// of the input indices.
-    fn get_many(&self, indices: &[Index]) -> Vec<T> {
-        fn sort_to_match_requested_index_order<T>(indexed_elements: HashMap<usize, T>) -> Vec<T> {
-            let mut elements = indexed_elements.into_iter().collect_vec();
-            elements.sort_unstable_by_key(|&(index_position, _)| index_position);
-            elements.into_iter().map(|(_, element)| element).collect()
-        }
-
-        // Do *not* refer to `self` after this point, instead use `self_lock`
-        assert!(
-            indices.iter().all(|x| *x < self.len()),
-            "Out-of-bounds. Got indices {indices:?} but length was {}. persisted vector name: {}",
-            self.len(),
-            self.name
-        );
-
-        let (indices_of_elements_in_cache, indices_of_elements_not_in_cache): (Vec<_>, Vec<_>) =
-            indices
-                .iter()
-                .copied()
-                .enumerate()
-                .partition(|&(_, index)| self.cache.contains_key(&index));
-
-        let mut fetched_elements = HashMap::with_capacity(indices.len());
-        for (index_position, index) in indices_of_elements_in_cache {
-            let value = self.cache.get(&index).unwrap().clone();
-            fetched_elements.insert(index_position, value);
-        }
-
-        let no_need_to_lock_database = indices_of_elements_not_in_cache.is_empty();
-        if no_need_to_lock_database {
-            return sort_to_match_requested_index_order(fetched_elements);
-        }
-
-        let keys_for_indices_not_in_cache = indices_of_elements_not_in_cache
-            .iter()
-            .map(|&(_, index)| self.get_index_key(index))
-            .collect_vec();
-        let elements_fetched_from_db = self
-            .reader
-            .get_many(&keys_for_indices_not_in_cache)
-            .into_iter()
-            .map(|x| x.unwrap().into());
-
-        let indexed_fetched_elements_from_db = indices_of_elements_not_in_cache
-            .iter()
-            .map(|&(index_position, _)| index_position)
-            .zip_eq(elements_fetched_from_db);
-        fetched_elements.extend(indexed_fetched_elements_from_db);
-
-        sort_to_match_requested_index_order(fetched_elements)
-    }
-
-    /// Return all stored elements in a vector, whose index matches the StorageVec's.
-    /// It's the caller's responsibility that there is enough memory to store all elements.
-    fn get_all(&self) -> Vec<T> {
-        let (indices_of_elements_in_cache, indices_of_elements_not_in_cache): (Vec<_>, Vec<_>) =
-            (0..self.len()).partition(|index| self.cache.contains_key(index));
-
-        let mut fetched_elements: Vec<Option<T>> = vec![None; self.len() as usize];
-        for index in indices_of_elements_in_cache {
-            let element = self.cache[&index].clone();
-            fetched_elements[index as usize] = Some(element);
-        }
-
-        let no_need_to_lock_database = indices_of_elements_not_in_cache.is_empty();
-        if no_need_to_lock_database {
-            return fetched_elements
-                .into_iter()
-                .map(|x| x.unwrap())
-                .collect_vec();
-        }
-
-        let keys = indices_of_elements_not_in_cache
-            .iter()
-            .map(|x| self.get_index_key(*x))
-            .collect_vec();
-        let elements_fetched_from_db = self
-            .reader
-            .get_many(&keys)
-            .into_iter()
-            .map(|x| x.unwrap().into());
-        let indexed_fetched_elements_from_db = indices_of_elements_not_in_cache
-            .into_iter()
-            .zip_eq(elements_fetched_from_db);
-        for (index, element) in indexed_fetched_elements_from_db {
-            fetched_elements[index as usize] = Some(element);
-        }
-
-        fetched_elements
-            .into_iter()
-            .map(|x| x.unwrap())
-            .collect_vec()
-    }
-
-    fn set(&mut self, index: Index, value: T) {
-        // Disallow setting values out-of-bounds
-
-        assert!(
-            index < self.len(),
-            "Out-of-bounds. Got {index} but length was {}. persisted vector name: {}",
-            self.len(),
-            self.name
-        );
-
-        self.write_op_overwrite(index, value);
-    }
-
-    /// set multiple elements.
-    ///
-    /// panics if key_vals contains an index not in the collection
-    ///
-    /// It is the caller's responsibility to ensure that index values are
-    /// unique.  If not, the last value with the same index will win.
-    /// For unordered collections such as HashMap, the behavior is undefined.
-    fn set_many(&mut self, key_vals: impl IntoIterator<Item = (Index, T)>) {
-        let self_len = self.len();
-
-        for (index, value) in key_vals.into_iter() {
-            assert!(
-                index < self_len,
-                "Out-of-bounds. Got {index} but length was {}. persisted vector name: {}",
-                self_len,
-                self.name
-            );
-
-            self.write_op_overwrite(index, value);
-        }
-    }
-
-    fn pop(&mut self) -> Option<T> {
-        // If vector is empty, return None
-        if self.is_empty() {
-            return None;
-        }
-
-        // add to write queue
-        self.write_queue.push_back(VecWriteOperation::Pop);
-
-        // Update length
-        *self.current_length.as_mut().unwrap() -= 1;
-
-        // try cache first
-        let current_length = self.len();
-        if self.cache.contains_key(&current_length) {
-            self.cache.remove(&current_length)
-        } else {
-            // then try persistent storage
-            let key = self.get_index_key(current_length);
-            self.reader.get(key).map(|value| value.into())
-        }
-    }
-
-    fn push(&mut self, value: T) {
-        // add to write queue
-        self.write_queue
-            .push_back(VecWriteOperation::Push(value.clone()));
-
-        // record in cache
-        let current_length = self.len();
-        let _old_val = self.cache.insert(current_length, value);
-
-        // note: we cannot naively remove any previous `Push` ops with
-        // this value from the write_queue (to reduce disk i/o) because
-        // there might be corresponding `Pop` op(s).
-
-        // update length
-        self.current_length = Some(current_length + 1);
-    }
-}
-
-impl<ParentKey, ParentValue, T> DbTable<ParentKey, ParentValue>
-    for DbtVecPrivate<ParentKey, ParentValue, Index, T>
-where
-    ParentKey: From<Index>,
-    ParentValue: From<T>,
-    T: Clone,
-    T: From<ParentValue>,
-    ParentKey: From<(ParentKey, ParentKey)>,
-    ParentKey: From<u8>,
-    Index: From<ParentValue>,
-    ParentValue: From<Index>,
-{
-    /// Collect all added elements that have not yet been persisted
-    fn pull_queue(&mut self) -> Vec<WriteOperation<ParentKey, ParentValue>> {
-        let maybe_original_length = self.persisted_length();
-        // necessary because we need maybe_original_length.is_none() later
-        let original_length = maybe_original_length.unwrap_or(0);
-        let mut length = original_length;
-        let mut queue = vec![];
-        while let Some(write_element) = self.write_queue.pop_front() {
-            match write_element {
-                VecWriteOperation::OverWrite((i, t)) => {
-                    let key = self.get_index_key(i);
-                    queue.push(WriteOperation::Write(key, Into::<ParentValue>::into(t)));
-                }
-                VecWriteOperation::Push(t) => {
-                    let key = self.get_index_key(length);
-                    length += 1;
-                    queue.push(WriteOperation::Write(key, Into::<ParentValue>::into(t)));
-                }
-                VecWriteOperation::Pop => {
-                    let key = self.get_index_key(length - 1);
-                    length -= 1;
-                    queue.push(WriteOperation::Delete(key));
-                }
-            };
-        }
-
-        if original_length != length || maybe_original_length.is_none() {
-            let key = Self::get_length_key(self.key_prefix);
-            queue.push(WriteOperation::Write(
-                key,
-                Into::<ParentValue>::into(length),
-            ));
-        }
-
-        self.cache.clear();
-
-        queue
-    }
-
-    fn restore_or_new(&mut self) {
-        if let Some(length) = self.reader.get(Self::get_length_key(self.key_prefix)) {
-            self.current_length = Some(length.into());
-        } else {
-            self.current_length = Some(0);
-        }
-        self.cache.clear();
-        self.write_queue.clear();
-    }
-}
-
-// possible future extension
-// pub struct DbtHashMap<Key, Value, K, V> {
-//     parent: Arc<RefCell<DbtSchema<Key, Value>>>,
-// }
-
-pub trait StorageSingleton<T>
-where
-    T: Clone,
-{
-    fn get(&self) -> T;
-    fn set(&mut self, t: T);
-}
-
-pub(crate) struct DbtSingletonPrivate<ParentKey, ParentValue, T> {
-    current_value: T,
-    old_value: T,
-    key: ParentKey,
-    reader: Arc<dyn StorageReader<ParentKey, ParentValue> + Sync + Send>,
-}
-
-pub struct DbtSingleton<ParentKey, ParentValue, T> {
-    // note: Arc is not needed, because we never hand out inner to anyone.
-    inner: RwLock<DbtSingletonPrivate<ParentKey, ParentValue, T>>,
-}
-
-impl<ParentKey, ParentValue, T> DbtSingleton<ParentKey, ParentValue, T>
-where
-    T: Default,
-{
-    // DbtSingleton can not be instantiated directly outside of this crate.
-    pub(crate) fn new(
-        key: ParentKey,
-        reader: Arc<dyn StorageReader<ParentKey, ParentValue> + Sync + Send>,
-    ) -> Self {
-        let singleton = DbtSingletonPrivate::<ParentKey, ParentValue, T> {
-            current_value: Default::default(),
-            old_value: Default::default(),
-            key,
-            reader,
-        };
-        Self {
-            inner: RwLock::new(singleton),
-        }
-    }
-
-    pub(crate) fn read_lock(
-        &self,
-    ) -> RwLockReadGuard<'_, DbtSingletonPrivate<ParentKey, ParentValue, T>> {
-        self.inner.read().unwrap()
-    }
-
-    pub(crate) fn write_lock(
-        &self,
-    ) -> RwLockWriteGuard<'_, DbtSingletonPrivate<ParentKey, ParentValue, T>> {
-        self.inner.write().unwrap()
-    }
-}
-
-impl<ParentKey, ParentValue, T> DbTable<ParentKey, ParentValue>
-    for DbtSingleton<ParentKey, ParentValue, T>
-where
-    T: Eq + Clone + Default + From<ParentValue>,
-    ParentValue: From<T> + Debug,
-    ParentKey: Clone,
-{
-    fn pull_queue(&mut self) -> Vec<WriteOperation<ParentKey, ParentValue>> {
-        self.write_lock().pull_queue()
-    }
-
-    fn restore_or_new(&mut self) {
-        self.write_lock().restore_or_new()
-    }
-}
-
-impl<ParentKey, ParentValue, T> DbTable<ParentKey, ParentValue>
-    for DbtSingletonReference<ParentKey, ParentValue, T>
-where
-    T: Eq + Clone + Default + From<ParentValue>,
-    ParentValue: From<T> + Debug,
-    ParentKey: Clone,
-{
-    fn pull_queue(&mut self) -> Vec<WriteOperation<ParentKey, ParentValue>> {
-        self.borrow_mut().pull_queue()
-    }
-
-    fn restore_or_new(&mut self) {
-        self.borrow_mut().restore_or_new()
-    }
-}
-
-impl<ParentKey, ParentValue, T> StorageSingleton<T> for DbtSingleton<ParentKey, ParentValue, T>
-where
-    T: Clone + From<ParentValue> + Default,
-{
-    fn get(&self) -> T {
-        self.read_lock().current_value.clone()
-    }
-
-    fn set(&mut self, t: T) {
-        self.write_lock().current_value = t;
-    }
-}
-
-impl<ParentKey, ParentValue, T> StorageSingleton<T>
-    for DbtSingletonReference<ParentKey, ParentValue, T>
-where
-    T: Clone + From<ParentValue> + Default,
-{
-    fn get(&self) -> T {
-        self.borrow().get()
-    }
-
-    fn set(&mut self, t: T) {
-        self.borrow_mut().set(t)
-    }
-}
-
-impl<ParentKey, ParentValue, T> StorageSingleton<T>
-    for DbtSingletonPrivate<ParentKey, ParentValue, T>
-where
-    T: Clone + From<ParentValue>,
-{
-    fn get(&self) -> T {
-        self.current_value.clone()
-    }
-
-    fn set(&mut self, t: T) {
-        self.current_value = t;
-    }
-}
-
-impl<ParentKey, ParentValue, T> DbTable<ParentKey, ParentValue>
-    for DbtSingletonPrivate<ParentKey, ParentValue, T>
-where
-    T: Eq + Clone + Default + From<ParentValue>,
-    ParentValue: From<T> + Debug,
-    ParentKey: Clone,
-{
-    fn pull_queue(&mut self) -> Vec<WriteOperation<ParentKey, ParentValue>> {
-        if self.current_value == self.old_value {
-            vec![]
-        } else {
-            self.old_value = self.current_value.clone();
-            vec![WriteOperation::Write(
-                self.key.clone(),
-                self.current_value.clone().into(),
-            )]
-        }
-    }
-
-    fn restore_or_new(&mut self) {
-        self.current_value = match self.reader.get(self.key.clone()) {
-            Some(value) => value.into(),
-            None => T::default(),
-        }
-    }
-}
-
-pub struct DbtSchema<ParentKey, ParentValue, Reader: StorageReader<ParentKey, ParentValue>> {
-    pub tables: Vec<Arc<RefCell<dyn DbTable<ParentKey, ParentValue> + Send + Sync>>>,
-    pub reader: Arc<Reader>,
-}
-
-impl<
-        ParentKey,
-        ParentValue,
-        Reader: StorageReader<ParentKey, ParentValue> + 'static + Sync + Send,
-    > DbtSchema<ParentKey, ParentValue, Reader>
-{
-    pub fn new_vec<I, T>(&mut self, name: &str) -> DbtVecReference<ParentKey, ParentValue, T>
-    where
-        ParentKey: From<Index> + 'static,
-        ParentValue: From<T> + 'static,
-        T: Clone + From<ParentValue> + 'static,
-        ParentKey: From<(ParentKey, ParentKey)>,
-        ParentKey: From<u8>,
-        Index: From<ParentValue>,
-        ParentValue: From<Index>,
-        Index: From<u64> + 'static,
-        DbtVec<ParentKey, ParentValue, Index, T>: DbTable<ParentKey, ParentValue> + Send + Sync,
-    {
-        assert!(self.tables.len() < 255);
-        let reader = self.reader.clone();
-        let key_prefix = self.tables.len() as u8;
-        let vector = DbtVec::<ParentKey, ParentValue, Index, T>::new(reader, key_prefix, name);
-
-        let arc_refcell_vector = Arc::new(RefCell::new(vector));
-        self.tables.push(arc_refcell_vector.clone());
-        arc_refcell_vector
-    }
-
-    // possible future extension
-    // fn new_hashmap<K, V>(&self) -> Arc<RefCell<DbtHashMap<K, V>>> { }
-
-    pub fn new_singleton<S>(
-        &mut self,
-        key: ParentKey,
-    ) -> DbtSingletonReference<ParentKey, ParentValue, S>
-    where
-        S: Default + Eq + Clone + 'static,
-        ParentKey: 'static,
-        ParentValue: From<S> + 'static,
-        ParentKey: From<(ParentKey, ParentKey)> + From<u8>,
-        DbtSingleton<ParentKey, ParentValue, S>: DbTable<ParentKey, ParentValue> + Send + Sync,
-    {
-        let singleton = DbtSingleton::<ParentKey, ParentValue, S>::new(key, self.reader.clone());
-        let arc_refcell_singleton = Arc::new(RefCell::new(singleton));
-        self.tables.push(arc_refcell_singleton.clone());
-        arc_refcell_singleton
-    }
-}
-
-pub trait StorageWriter<ParentKey, ParentValue> {
-    fn persist(&mut self);
-    fn restore_or_new(&mut self);
-}
-
-#[derive(Clone, PartialEq, Eq)]
-pub struct RustyKey(pub Vec<u8>);
-impl From<u8> for RustyKey {
-    fn from(value: u8) -> Self {
-        Self([value].to_vec())
-    }
-}
-impl From<(RustyKey, RustyKey)> for RustyKey {
-    fn from(value: (RustyKey, RustyKey)) -> Self {
-        let v0 = value.0 .0;
-        let v1 = value.1 .0;
-        RustyKey([v0, v1].concat())
-    }
-}
-impl From<u64> for RustyKey {
-    fn from(value: u64) -> Self {
-        RustyKey(value.to_be_bytes().to_vec())
-    }
-}
-
-#[derive(Debug)]
-pub struct RustyValue(pub Vec<u8>);
-
-impl From<Vec<u8>> for RustyValue {
-    fn from(value: Vec<u8>) -> Self {
-        RustyValue(value)
-    }
-}
-impl From<RustyValue> for Vec<u8> {
-    fn from(value: RustyValue) -> Self {
-        value.0
-    }
-}
-impl From<RustyValue> for u64 {
-    fn from(value: RustyValue) -> Self {
-        u64::from_be_bytes(value.0.try_into().unwrap())
-    }
-}
-impl From<u64> for RustyValue {
-    fn from(value: u64) -> Self {
-        RustyValue(value.to_be_bytes().to_vec())
-    }
-}
-impl From<RustyValue> for crate::shared_math::tip5::Digest {
-    fn from(value: RustyValue) -> Self {
-        crate::shared_math::tip5::Digest::new(
-            value
-                .0
-                .chunks(8)
-                .map(|ch| {
-                    u64::from_be_bytes(ch.try_into().expect("Cannot cast RustyValue into Digest"))
-                })
-                .map(BFieldElement::new)
-                .collect::<Vec<_>>()
-                .try_into().expect("Can cast RustyValue into BFieldElements but number does not match that of Digest."),
-        )
-    }
-}
-impl From<crate::shared_math::tip5::Digest> for RustyValue {
-    fn from(value: crate::shared_math::tip5::Digest) -> Self {
-        RustyValue(
-            value
-                .values()
-                .map(|b| b.value())
-                .map(u64::to_be_bytes)
-                .concat(),
-        )
-    }
-}
-
-pub struct RustyReader {
-    db: DB,
-}
-
-impl StorageReader<RustyKey, RustyValue> for RustyReader {
-    fn get(&self, key: RustyKey) -> Option<RustyValue> {
-        self.db
-            .get(&ReadOptions::new(), &key.0)
-            .unwrap()
-            .map(RustyValue)
-    }
-
-    fn get_many(&self, keys: &[RustyKey]) -> Vec<Option<RustyValue>> {
-        let mut res = vec![];
-        for key in keys {
-            res.push(
-                self.db
-                    .get(&ReadOptions::new(), &key.0)
-                    .unwrap()
-                    .map(RustyValue),
-            );
-        }
-
-        res
-    }
-}
-
-/// Database schema and tables logic for RustyLevelDB. You probably
-/// want to implement your own storage class after this example so
-/// that you can hardcode the schema in new(). But it is nevertheless
-/// possible to use this struct and add to the scheme after calling
-/// new() (that's what the tests do).
-pub struct SimpleRustyStorage {
-    // pub db: Arc<DB>,
-    pub schema: DbtSchema<RustyKey, RustyValue, SimpleRustyReader>,
-}
-
-impl StorageWriter<RustyKey, RustyValue> for SimpleRustyStorage {
-    fn persist(&mut self) {
-        let write_batch = WriteBatch::new();
-        for table in &self.schema.tables {
-            let operations = table.borrow_mut().pull_queue();
-            for op in operations {
-                match op {
-                    WriteOperation::Write(key, value) => write_batch.put(&key.0, &value.0),
-                    WriteOperation::Delete(key) => write_batch.delete(&key.0),
-                }
-            }
-        }
-
-        self.db()
-            .write(&WriteOptions::new(), &write_batch)
-            .expect("Could not persist to database.");
-    }
-
-    fn restore_or_new(&mut self) {
-        for table in &self.schema.tables {
-            table.borrow_mut().restore_or_new();
-        }
-    }
-}
-
-impl SimpleRustyStorage {
-    pub fn new(db: DB) -> Self {
-        let schema = DbtSchema::<RustyKey, RustyValue, SimpleRustyReader> {
-            tables: Vec::new(),
-            reader: Arc::new(SimpleRustyReader { db }),
-        };
-        Self { schema }
-    }
-
-    #[inline]
-    pub(crate) fn db(&self) -> &DB {
-        &self.schema.reader.db
-    }
-}
-
-pub struct SimpleRustyReader {
-    db: DB,
-}
-
-impl StorageReader<RustyKey, RustyValue> for SimpleRustyReader {
-    fn get(&self, key: RustyKey) -> Option<RustyValue> {
-        self.db
-            .get(&ReadOptions::new(), &key.0)
-            .unwrap()
-            .map(RustyValue)
-    }
-
-    fn get_many(&self, keys: &[RustyKey]) -> Vec<Option<RustyValue>> {
-        keys.iter()
-            .map(|key| {
-                self.db
-                    .get(&ReadOptions::new(), &key.0)
-                    .unwrap()
-                    .map(RustyValue)
-            })
-            .collect()
-    }
-}
+mod dbtsingleton;
+mod dbtsingleton_private;
+mod dbtsingleton_ref;
+mod dbtvec;
+mod dbtvec_private;
+mod dbtvec_ref;
+mod enums;
+mod rusty_key;
+mod rusty_reader;
+mod rusty_value;
+mod schema;
+mod simple_rusty_reader;
+mod simple_rusty_storage;
+mod traits;
+
+pub use dbtsingleton::*;
+// pub use dbtsingleton_private::*;
+pub use dbtsingleton_ref::*;
+pub use dbtvec::*;
+pub use dbtvec_private::*;
+pub use dbtvec_ref::*;
+pub use enums::*;
+pub use rusty_key::*;
+pub use rusty_reader::*;
+pub use rusty_value::*;
+pub use schema::*;
+pub use simple_rusty_reader::*;
+pub use simple_rusty_storage::*;
+pub use traits::*;
 
 #[cfg(test)]
 mod tests {
 
+    use std::sync::Arc;
+
     use rand::{random, Rng, RngCore};
 
-    use crate::shared_math::other::random_elements;
+    use crate::{
+        shared_math::other::random_elements,
+        storage::{
+            level_db::DB,
+            storage_vec::{Index, StorageVec},
+        },
+    };
+    use itertools::Itertools;
 
     use super::*;
